feat(api,web,shared-types): add Resend email provider, ARIA labels, XOF validation, payment visibility
- Add Resend as default email provider with tags, batch sending, scheduled
  sends, idempotency keys, and attachment support (10 tests)
- Keep SendGrid as fallback in provider registry
- Add French aria-label to 26 icon-only buttons across backoffice and
  participant apps for accessibility compliance
- Enforce XOF integer validation: .int() on ticket prices, refine promo
  code fixed discounts, runtime check on refund amounts
- Show payment_failed status in registration list, failed payment count
  in summary cards, failure reason inline in payment table
- Add pending/expired filters to payment status dropdown
- Update delivery plan docs to reflect Resend integration (337 tests)

+++ packages/shared-types/src/event.types.ts
@@ -128,7 +128,7 @@ export const TicketTypeSchema = z.object({
   id: z.string(),
   name: z.string(),                        // e.g. "VIP", "Standard", "Press"
   description: z.string().optional(),
-  price: z.number().min(0).default(0),     // 0 = free
+  price: z.number().int().min(0).default(0), // XOF — no decimals, 0 = free
   currency: z.enum(["XOF", "EUR", "USD"]).default("XOF"), // XOF = CFA Franc
   totalQuantity: z.number().int().positive().nullable(), // null = unlimited
   soldCount: z.number().int().default(0),
